Add update links and portrait mod.
I keep forgetting the basic stuff and I don't want to amend the commit,
so here's another commit with the update keys added.  Also took the
opportunity to add the content patcher mod I created with Einari's
portraits with other fixes as noted in the manifest file.  This is an
internal mod and will probably not be released on nexus since the
original mod authors have already released their versions, but I'll make
public for anyone who likes this amalgammation of art.